refactor(realtime-v3): split oversized Realtime/Channel into extension files
Both Realtime.swift (~1150) and Channel.swift (~1050) were over the 800-line
house soft-max. Carve out cohesive extensions (no behavior change):

- Realtime+Push.swift      — nextRef/_rawSend/_push/_encodeToJSON/awaitReply/accessTokenForJoin
- Realtime+Reconnection.swift — handleConnectionLost + reconnection loop + rejoin/give-up
- Realtime+Heartbeat.swift  — startHeartbeat/_sendHeartbeat/updateLatency
- Channel+Routing.swift     — receive + server-close/error + system routing + server-id routing
- Channel+Presence.swift    — gains the presence track/untrack send seam

Stored properties stay in the primary declarations; members reached across the
new file boundaries are widened from private to internal (Swift private does not
cross files), each annotated with why. Realtime.swift → 778, Channel.swift → 786.
